Move string validation to caller
